fastapi-oauth/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pymysql==1.1.1
python-dotenv==1.0.1
Authlib==1.3.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.24.1
pydantic[email]==2.8.0
itsdangerous==2.2.0
fastapi
firebase-admin